refactor: use a third party package for email valildation
