+++ project-words.txt
@@ -15,11 +15,13 @@ Cyberneering
 datetime
 DATETIME
 Dont
+dotless
 Grünwald
 hasher
 Hasher
 hexlify
 httpseeds
+ICANN
 imagoodboy
 imdl
 indexadmin
