+++ src/services/user.rs
@@ -13,7 +13,7 @@ use crate::errors::ServiceError;
 use crate::mailer;
 use crate::mailer::VerifyClaims;
 use crate::models::user::{UserCompact, UserId, UserProfile};
-use crate::utils::regex::validate_email_address;
+use crate::utils::validation::validate_email_address;
 use crate::web::api::v1::contexts::user::forms::RegistrationForm;
 
 /// Since user email could be optional, we need a way to represent "no email"
