+++ src/utils/mod.rs
@@ -1,4 +1,4 @@
 pub mod clock;
 pub mod hex;
 pub mod parse_torrent;
-pub mod regex;
+pub mod validation;
